add fp32/fp64 global config
The floating point precision is now determined by the environment variable SKSHAPES_FLOAT_DTYPE possible values are float32 and float64 with float32 by default
Tests are run twice during CI : with float32 and with float64 and assert of dtypes were added to check persistence of floating point type after skshapes transformations

+++ skshapes/tasks/registration.py
@@ -1,7 +1,7 @@
 """Registration between two shapes."""
 
 from ..optimization import Optimizer
-from ..types import typecheck, Union, shape_type
+from ..types import typecheck, Union, shape_type, float_dtype
 from ..loss import Loss
 from ..morphing import Model
 import torch
@@ -79,7 +79,9 @@ class Registration:
 
         # Initialize the parameter tensor using the template provided by the model, and set it to be optimized
         parameter_shape = self.model.parameter_shape(shape=source)
-        parameter = torch.zeros(parameter_shape, device=self.optim_device)
+        parameter = torch.zeros(
+            parameter_shape, device=self.optim_device, dtype=float_dtype
+        )
         parameter.requires_grad = True
 
         # Initialize the optimizer
